+++ projects/ids-enterprise-ng/src/lib/datagrid/index.ts
@@ -1,7 +1,6 @@
 export {
   SohoDataGridComponent,
   SohoDataGridToggleRowEvent,
-  SohoGridColumnFilterTypes,
   SohoAngularEditorAdapter
 } from './soho-datagrid.component';
 export { SohoDataGridModule } from './soho-datagrid.module';